Base Player Prefab
Creating the base player prefab. Refer to the player prefab meeting to see initial architecture that was decided. This is just setting up what was decided within the meeting. All of the scripts are empty and the values of the components are at their defaults. All future work on player prefab must start off this one.