Added item prefab and made it droppable
Created a very basic item prefab. This object is what will appear
throughout the game. The only information it contains is the ID of the
item that it represents and the corresponding icon for that item.

Changelog:
* Added Item prefab
* Updated DropManager: it has the ability to make any given item
appear in the world (dropping it)
* Updated ItemManager: making sure it provided all of the necessary
information to the DropManager
* Added a script for the item game object that stores the item's ID but
also takes care of setting up the game object