Preparations for the new passive items
Modified the various player scripts and the base item scripts in
preparation for adding all of the new items. When a player picks up an
item, the item's effect will always be applied immediately onto the
PlayerManager script. At this point, the manager and the item decide how
the effect works and depending on the intention, the item is sent in
different directions.<br>
Added temporary sprites for the first five passive items as well as a
sprite to represent any general item.

Changelog:
* BaseItem/BasePassiveItem: some new variables and an updated interface
that requires a player manager. All items talk to the PlayerManager
first.
* PlayerWeapon/WeaponBase: a list that contains functions describing any
effect that must occur when the weapon hits a target and an execution of
that list.
* PlayerManager/PlayerData/Inventory/UnitStats: making sure that items
can be picked up and dropped properly. When an item gets picked up, it
disappears and it reappears when it is dropped. Any item in the
inventory has its effect applied. Lots of different functions to enable
items to talk to the various components and attributes of the player.
* Added a backpack sprite to represent any general item
* Added sprites for:
  * AngrySun
  * HeartyHeart
  * MagnifyingGlass
  * RunningShoes
  * TiredHeart