Adding actives
The main purpose of this commit is to add the three actives: Surrender,
Chowtime, and BowNotIncl.<br>

Changelog:
- PlayerManager
  - Stores a reference to the aim direction. Defaults to
(1,0) because the function that uses it cannot accept a (0,0) direction.
It needs to know where to fire the projectile. The function being
referred to is part of the BowNotIncl active item.
  - Can activate temporary effects both for general stuff that would be
  permanent otherwise and for on attack effects
  - We might want to redesign how these effects get applied because it
  may become unwieldy
- Inventory cycling
  - Players can now perform very simple cycling. The event stores the
  direction and the update function takes care of cycling through the
  inventory if relevant. There is a very minor delay between each "next"
  item to make going to the neighbor easier. Alternatively, it would be
  too easy to overshoot and have trouble selecting a particular item
- BaseActiveItem has been slightly updated so it provides read access to
some key variables
- Update of UnitStats to use a dictionary for the status effects. Also,
the "invulnerable" condition has been added to the dictionary. As we add
more status effects into the game, we should move everything into the
dictionary as it cleans everything up and makes the code more concise
and generally more scalable.
- Change various scripts and variables to use m_ as opposed to p_, cc_,
and cr_
- Added the three active items: Surrender, Chowtim, and BowNotIncl
- BowNotIncl requires a projectile so a general prefab and a
corresponding script has been prepared