Cleaning up Inventory and ItemManager
The ItemManager and Inventory scripts became somewhat confusing and
unwieldy. This fixes them up, removing unnecessary functions, renaming
functions, and adding more generics to reduce repeated code.
Additionally, the two constants' scripts have been merged into Consts.

Changelog:
- ItemManager is completely static. Nearly all of the methods in this class are
static and based on a singleton. When a script asks for information on
an item, the ItemManager will first check whether or not an instance of
it exists and if it does, it will use that singleton instance to return
the necessary information.
- Inventory has been cleaned up. A lot of methods have been removed and
some renamed. Interface methods have had the amount of work they perform
reduced and outsourced to other methods called by them. Logically,
everything should be the same but the script should be more readable
now.
- Consts and Constants have been merged into Consts