Adding player basic attack. Daggers + polearms
Description:
The main purpose of this commit is to add functionality for the player
to attack using any melee weapon they pick up. When the player picks up
a weapon, it will be equipped. Once equipped, the weapon rotates around
the player in a 360 degree circle based on the direction that the player
has the right trigger facing. When attacking, the weapon will start with
a wind up followed by the actual attack animation. When attacking, a
series of raycasts will be shot forward in the same arc of the weapon
itself*. If a raycast hits another player then damage is dealt and the
other player is pushed back. Only one raycast can be applied per player
per attack so in order to hit the same player twice, they must be
attacked two separate times. Immediately before damage and knockback
take their effect, any item effects that should be applied do get
applied and modify the two stats. By default, weapons get used for as
long as the attack button is held. Once the attack animation ends, the
weapon goes back to its original position and waits there until it is
needed to attack again.
*weapons can attack in an arc or a stabbing motion.

Testing:
1. Go into any scene that is set up to spawn players
2. Add some weapon items and passive/active items that affect weapon
stats
   - Ketchup Dagger ID: 6667
   - Mustard Dagger ID: 6668
   - The Basic Spear ID: 6669
   - Angry Sun (passive) ID: 5
     - Increases knockback by a lot
   - Chowtime (active) ID: 3335
     - Increases damage by a lot for a few seconds
3. Start the game. Make sure to spawn TWO or more players
4. Have one or both of the players pick up a weapon (items are optional)
5. Hold the attack button for some time while close to the other players
6. Notice the other players dying because they still need to git gud

Important Changes:
- Players can now attack when they pick up a weapon.
- Player prefab no longer has the sound trigger because it conflicts
with the basic attack logic. Once we start implementing SFX, this should
be tackled.
- Set the layers that are hittable in GameManager

Changelog:
- A weapon is picked up only when: there is space OR it is different
from the currently equipped weapon
- Updated player prefab so its size is (1,1,1) as opposed to (3,3,1). To
accomodate this, the sprites have all had their pixel per unity reduced
from 100 to 30.
- Updated PlayerWeapon so that it moves the weapon around the player in
a 360 degree circle
  - The weapon is place at a slight offset and can be given an angle
  offset. These two allow clean positioning of the weapon at all times
- The weapon attack is based on four stages:
  1. Windup
  2. Attack
  3. Reset to original position
  4. Wait until ready to attack again
- The total length of time between attacks is equal to 1 / attack speed.
Each of the four stages is divided by using percentages that are hard
coded per weapon type. Windup for daggers takes the first 10% of the
time, attack animation the next 30%, and so on.
- When the player spawns, the game manager sets the physics layer for
the player to something different from the other players so that when
the weapon is used, it does not hit the player that used it.
- Weapons can be set to attack in an arc or in a stabbing motion. For
the arc, the position vectors are irrelevant but for the stabbing, arc
angle is still important. Whether a weapon stabs or attacks in an arc is
currently based on the actual weapon rather than the type
- During the attack stage, a series of raycasts are shot out following
the arc of the weapon. Each raycast overlaps all enemies within its
bounds. The enemies hit are kept track of so that for weapons with high
resolution, they don't deal extra damage.
- Weapons can only be switched when not attacking
- Weapons can only be dropped when not attacking (if inventory is open,
assumes not attacking)
- In the game manager, we set all of the layers that are hittable. The
hittable layers should be each of the player layers as well as the enemy
layer and potentially "destructible" item layer if it will exist. Once
set, all attacks by a player will only hit objects on those layers
EXCLUDING the layer that the attacker is on. This allows us to specify
which layers players can hit and at the same time ensure that they are
not hitting themselves.
- Added two daggers:
  - Ketchup Dagger (complete garbage)
  - Mustard Dagger (not complete garbage)
- Added polearm:
  - The Basic Spear (pretty OP)
- Players can have no weapon equipped. If a weapon is dropped, it does
not automatically switch to an equipped weapon. The player will have no
weapon equipped until they decide to pick one up or swap.

Bug Fixes:
- N/A

New Bugs:
- Not exactly a new bug but it is not noted anywhere: the BowNotIncl
active item (ID: 3336), currently deals no damage.
- Move/idle graphics don't work with weapon so the animator component
has been disabled and the "UpdateGFX" function in player movement has
been commented out