Combat UI (#83)
All available combat UI elements have been added in. As of right now
that means that weapons and active items are currently visible to the
player. CombatParent was added to the PlayerHUD which currently holds
the gems and images for Class Ability, ActiveItem, Weapon1 and Weapon2.

Changelog:
-Added CombatParent to PlayerHUD under the player prefab
-Added all the subcomponent gems and images (see description)
-Added a HealthUI section in PlayerCanvas
-Shifted Health related UI into HealthUI
-Added Swap/Drop/Set integration for Active item and Weapons
-Anchored the UI elements to support multiple players.

Testing:
-Make sure to set the combatParent in playercanvas script in player
prefab if it hasn't already been set.
-Nav over to WeyScene scene.
-Have multiple players ready to go.
-Pick up weapons, drop weapons, swap weapons
-Pick up Active items
-Open inventory and similarly drop the selected items.
-Fiddle around with any combination of picking up and dropping

Warnings:
-Class ability gem never gets populated b/c it doesn't exist.
-No gold or keys art/integration exists
-Visual scaling and adjusting of items in the gems hasn't been done yet.
-Thinking about waiting to meet up with an artist to fix those up b/c
none of it is code related, just adjusting size position and scale.