Buzzer Graphics Integrated (#127)
Changelog:
The buzzer is now an adorable dragon
He can flap his wings!
He flaps fast when he moves!
He'll turn to face the player (there is an actual turn in the animation which shows its a 2D sprite. I wasn't sure how to fix it, but I think it actually looks okay)
Cleaned up folders a bit (removed all the sample enemy animations, whatever those were)

Important Changes:
Facing direction is now updated in FindClosestTarget as opposed to Update, and is based off of the position of the closest player as opposed to the movement direction of the enemy.